Started an extended control version (not working yet)+ code clean up
New version of the code for the original example features a preparation for an additional state. However, after careful checking the agent only manages to learn a small portion of the time when running different seeds. 
Changed the way the policy distribution was evaluated (now everything is done within GradientTape and tfp distribution is used to generate log probs and samples).
Need more work to make it functional (check Gradients, and do parameter sweeping)