Enriched quantum environment for gate calibration procedure
QuantumEnvironment now does have a second method enabling the calibration of a target gate and not only prepare a target state. An example file is provided for this use case. Gate calibration procedure not working for now, need to test and debug. 
Other changes: Repo structure (now init file present for package loading)