Enhanced speed of simulation and made it compatible with Runtime
QuantumEnvironment now works correctly with all Runtime backends online, moreover calculation of expectation values is done faster now thanks to a regrouping of all the observables to be sampled. Main issue before was that the original quantum circuit contained a classical register, whereas Runtime needs to add it itself to do the expectation value calculation.
Tried an alternative way of handling the Runtime Session in script for arbitrary state prep, not relevant when dealing with qiskit.primitives.
To be tested next: qiskit_aer.primitives for noise simulation.