Forward with pulse level implementation
Introduced BackendEstimator and Qiskit Dynamics modules to run the algorithm while simulating from a real backend the output of the parametrized schedule/quantum circuit.
At the moment, problem with parametrization and/or transpilation to run the circuit, need to clarify this.
Still lacking, benchmarking based on the solver extracted from the class for gate fidelity estimation