Removed dependencies on deprecated qiskit opflow
QuantumEnvironment now can handle the input states generation by itself (based on Qiskit Experiments) and the user does not have to provide it in the Qiskit_config argument anymore (He still can, but is no longer required).
The code now also enables the calibration of a gate within a wider quantum circuit as long as the target "register" argument is provided and is not as large as the number of qubits.