Inserted QiskitConfig data class in workflow
All notebooks and python files now take a QiskitConfig data class instead of the the usual dictionary to store options. The field target_type in the target has also been removed.
Circuit context alaysis inititated through CircuitInstruction in QuantumCircuit.data? Unfortunately, we need scheduling info to enable the adding of parallel gates in the device. To be done at the schedule level then