Added step and reset methods to TF_Environment
Revised the QuantumEnvironment to handle new Runtime change on the Session handling (it internally opens a session).
The TF Environment now is able to run the step function as expected for the different circuit truncations. More will follow in the README to explain the functioning of the class once ready