Added ProcessTomography benchmark for average gate fidelity
For backends that are not from Runtime, one has the workflow available to perform everything as if one had access to real experiment.
Issue remaining for Runtime backends: Qiskit Experiments cannot be run within a Session (backend.run not being defined).
One should try to build a circuit_runner program within the session from the experiment workflow and deal with the data directly.