Added compatibility with backends carrying CX as basis gate
Everything gets down to a compilation to ECR gate at the moment since this is the proposed two-qubit gate present in the BasisGateLibrary. If one can build a custom template for his gate, one can replace the transpilation procedure. At this point, we force everything to be going through the ECR gate.