Revamped QuantumEnvironment declaration
All the hyperparameters and user defined variables should now be contained in a TrainingConfig dataclass, used to define all common parts for QuantumEnvironment and ContextAwareQuantumEnvironment (formerly TorchQuantumEnvironment). The idea is to enable the user to only touch a user file and leave the rest untouched throughout the entire computation.

What's next: integrate user file, and automated workflows inherent to training and hyperparameter optimization