Added backwards compatibility between Envs and validated learning
The ContextAwareQuantumEnvironment at the gate level also works for the same CX gate (although normalization is not yet aligned) as the original QuantumEnvironment. To be checked next at the pulse level. Moreover, compatibility of the PPO has been added to handle RescaleAction.
However, training with cal check does not yet work with the new env (mismatch between computations and initializations of new env)