Refactored pulse config and added generic Hamiltonian generator
To make the user config file for the pulse level easier, some functions were moved to the helper functions. That way, the user can focus on the minimal changes from one backend to another.
There is now in dynamics_config.py a new function enabling the generation of a DynamicsBackend taking arbitrary numbers of qubits. The terms are those depicted in the DynamicsBackend tutorials, but scaled up for any qubit connectivity. This should make the test on multi qubit devices easier.
This function could be further customised in future to account for more contextual effects.