Channel_Estimator new class option
Preparing a new way of calculating the reward to assess directly the gate fidelity and not state fidelity per epoch. There is now a new class attribute to QuantumEnvironment that can be set in the yaml file to choose which mode shall be selected (note that channel estimation is not working yet).
Other changes:
- Pulse benchmark now also works for DynamicsBackend without JaxSolver (however not very efficient as independent simulations are done, could use parameterized versions for more efficiency).
- Removed perform_action redefinition in context aware environment (the step function now calls the base class method, and computes the right parameters)
- Added some comments in some functions