Removed observation space as a user input
The observation space should not be specified anymore by the user, as it will be a key part of the protocol construction that we might want to keep control of. Configuration files have been updated accordingly.

Other features:
-Check Reward function now adjusts the c factor based on the fidelity obtained, which should therefore enable the superposition between the fidelity and reward curve systematically.
- Tried to change internally the observation space to include information about the observable to be computed with the estimator. Surprisingly, feeding this information does not improve learning but makes it slower (the way it was added was through coefficients of the Pauli terms, each stored in one input neuron, d**2 neurons being added to the two already existing.

Channel estimation for the ContextAwareQuantumEnvironment has been removed, as the characteristic function for the channel becomes way too large for already 5 qubits.