Updated code for new versions of Qiskit Dynamics and Runtime
Code has been updated to leverage behind the scenes the new Qiskit Runtime Local Service, enabling the seamless usage of different Estimator primitives (Aer, Backend, or Runtime) based on the actual backend provided. This enables a simplification of the choice of primitive previously done in the instanciation of a QuantumEnvironment.

Another update was done through the update of Qiskit Dynamics, removing the need to use the Array extension for the use of Jax.

Also adjusted the PPO file to plot the superposed curves for Fidelity and Reward no matter what the benchmark cycle.

Adjusted requirements.