Refactored Channel estimation to enable arbitrary number of input states
The channel estimation part is better handled through a division of the shots budget through the preparation of various input states (for each Pauli sampled as an input state). This should exemplify the ideal tradeoff required for a channel estimation task (hard to achieve if no real-time processing). There is also the possibility to adjust the number of input states that should be sampled (for now hardcoded within the method only, if relevant could become a hyperparameter in the future, to be tested first).

Small refactors are done elsewhere