Introducing QuantumEnvironmentV2
This commit revamps in a major way the abstractions behind the various QuantumEnvironments.

We now have a `BaseQuantumEnvironment`class that is used as a joint resource for both `QuantumEnvironment`and `ContextAwareQuantumEnvironment`.
That way, we avoid inconsistencies due to target and circuits redefinitions when a context-aware version was created (it had to first create the `QuantumEnvironment`and then rewrite those attributes.

The functionalities are overall the same, but a few changes in the way we create the config files are now used.

To give more flexibility for future developments, we use dedicated Options dataclasses for each subset of features describing the environment (namely how is the reward computed, or how is the fidelity benchmarking performed, or simply what hyperparameters do define the RL workflow).
This paves the way for more modularity, subclassing the environment for more complex workflows (e.g. an upcoming QUA dedicated version), while doing all the heavylifting in the baseclass using advanced Qiskit features.
The pulse level now successfully runs on IBM devices, and the workflow is ready to be tested on such architecture (if we get access to it again...).

At the moment, the notebooks are just adding a V2 version of the environments to make sure that all the files from other branches can be synchronized, but it should be fairly straightforward to adapt them.