Corrected bug using fidelity reward
Reward methods are now set in the init, and total shots and hardware runtime are naively updated in the case of a fidelity reward.
Added in comment of the ContextAware version the possibility of doing optimization over fidelity over subsystem of interest (could be useful to cross check validity of other rewards in the future).