Updated Target and benchmark function in QEnv
The way the layout and the target were defined in the base class was refactored to really indicate that one Target should contain one Layout only. Instead, the ContextAwareQuantumEnvironment class now carries an array of target, that are selected based on the trunc index of interest, leaving the functionality identical to our previous way of working.
Moreover, the compute benchmark function was adjusted to include both state (circuit) and gate fidelities for the context aware case.

Requirements are also updated