Updated gate calibration mechanism and benchmarking method
For pulse level,
We implement for the QuantumEnvironment a method `update_caibration()` enabling the attachment of the pulse schedule with optimal actions to the target gate in the Backend target.

Added further checks to add more flexibility in the building of QEnvConfig.

Corrected phase rotations to retrieve best unitary.

Refactored way to benchmark fidelity vs other reward method.

Added ECR notebook use case