Added coherent error script and added TransformationPass for CAEnv
This new pass enables the replacement of the original target instructions by a set of custom parametrized instructions. The filtering over nearest neighbors is not done yet and will be the subject of another pass. This pass will be useful to compute easily the new circuits that contain the custom operations as gates instead of the original gates.