Modified pulse level sim functions and changed backend loading method
Updated tomography and sim to support more generic workflows.
Updated also the notebook structures, where the backend is now declared within the notebook and can be fed directly to the get_q_env_config method.
Previous way of loading the callable is still doable, but now it does not have to necessarily take the redundant backend parameters for handling the yaml file (it is instead handled internally).