New version of context aware environment released in main file
The new environment is fully backwards compatible and recovers previous learning on easy circuits. It is now to be tested on newest spillover use case.

One thing to note: agent or environment wrapper has to be adapted to fully explore action space as rescale action wrapper was changed in Gymnasium 1.0