Major refactor of PPO
PPO class has been refactored for clarity and is now compatible with wandb to log data.

This refactoring was motivated by the fact that we only rely on this class approach to build the scripts, making the separation of functions not really. necessary and making it less clear.

The logging is handled in a separate module now, as well as the initialization.