Refactored rewards into separate files + adjusted QUA + edited Channel
Channel Reward is refactored, as well as StateReward to also provide indices that will be loaded into the QUA program later on.
The calculations are also redone although it does not yield good behavior yet.
Also proposed an extension (not ready) to deal with multiple n-reps per epoch of training). This can be tuned through the execution_config n_reps_mode parameter (joint or sequential)