Introducing General learning environment for local spillover noise
This fix tunes many enviroment components to enable the successful binding of a complex circuit dependent noise model, that is reset every time in new custom environments for enabling general learning of circuits composed of a single layer of single qubit rotations followed by a layer of two qubit gates.
Components related to the local noise are extremely fragile and should be modified with extreme care. The main components of the environment however should be working in the general case and remain fully functional (no fixes needed at the expense of losing generality, the custom environments do take care of that).