First step towards changing multiple circuit context handling
Remaining: ContextAwareQuantumEnvironment handling

This branch aims to refactor largely the computation of the circuit context handling, which in the past relied on truncations of one single circuit context. However the project seems to now focus more on contexts that are manually assigned by the user, rendering the truncation notion impractical.
This mechanism opens the door for disentangling the target definition from the custom circuits, as it's of the responsibility of the user now to specify within the Target object (to be directly created now, the config object related to it is pending removal) the set of contexts he'd like to assign. Then the reset function of the environment would be in charge of computing the circuit index to go for (the former trunc index becomes now circuit choice), and will have an impact on the next step of training.