New handling of circuit contexts within Target
The Target objects are now the primary interfaces which contain the specifications of the contexts in which the gates are meant to be executed in. This has the property to now enable the clear separation between the target and the circuit definitions from one single context.
From now on, the set of circuit contexts are to be specified by the user as a list, and he can choose which circuit context to use during training with the circuit_choice change attribute.

We removed the pass based on the truncations of one single circuit, as this was not exhibiting scalable workflows. We will leave commented out the previous computation in case one would add two parameter sets for the same gate appearing twice in the circuit. It would probably not be too complex to adapt it and extend the circuits definition but that should not be the default mode.