Refactors QUA environment for RL-QOC
This commit introduces significant refactoring to the QUA environment to facilitate reinforcement learning-based quantum optimal control (RL-QOC). It introduces a new `CircuitParams` dataclass for managing circuit parameters within QUA programs. This change modularizes parameter handling, and allows for more complex reward schemes.

The environment's step function is streamlined for improved efficiency in pushing and fetching data to and from the OPX. Reward computation is also refactored to work with the different reward methods, state, channel, and cafe, to execute a QUA program designed specifically for the RL-based calibration project and handle the data streaming and processing.

This commit also introduces a test mode to save all the parameters streamed to OPX.

The hope is to get custom QUA programs for each reward method that is introduced. One could think about introducing BenchmarkConfig as an extra argument and specify there a custom QUA macro to be leveraged. Left for future work