New observation space handling +Added parametric rx gate use case
TODO: Rewrite SyncHook file to contain all necessary Gym wrappers for this use case

The observation space has now been complexified whenever the provided target gate is parametrized by default. This will enable the calibration of a parametric gate with the ContextAwareQuantumEnvironment, provided that it is wrapped up in the new ParametricGateWrapper.