Refactors sync hook generation
Updates sync hook generation to directly utilize the environment,
simplifying the process and enhancing code maintainability.

This change replaces the manual passing of target, reward, and other
environment parameters with direct environment access.

Furthermore, it generalizes the usage of different Gymnasium wrappers wrapping the BaseQuantumEnvironment to recreate a similar pattern of interaction between the agent and the environment on the IQCC side