Generalized InstructionReplacement and first trial for MultiTarget
A major internal refactoring of the InstructionReplacement and GateTarget were made in order to prepare the ground for a new multi-agent environment that will be capable of calibrating multiple gates in parallel. For this, the first step was to move from a target disconnected from its practical implementation to a unified component that indicates the target instruction and its custom replacement to be integrated within a wider circuit context.
In the future, we will move away from the existing callable method that was previously used to store the information about the custom replacement. Instead, we will focus on putting a QuantumCircuit (for circuit simulation) or a custom Instruction (pulse level) with some additional InstructionProperties that will be acting as replacers. We have changed the internals of the currently existing environments to also accommodate for this slight change of the InstructionReplacement and CustomGateReplacementPass. The existing codebase should remain valid, but future implementations will involve another more scalable pipeline.